refactor: more rich column to own file
